pyproject.toml
[project]
name = "face-generation"
version = "0.1.0"
description = "Face generation application whith ability to train VAE and DCGAN."
authors = [
    {name = "Ivan Konyushenko",email = "idkonyushenko@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dvc (>=3.60.0,<4.0.0)",
    "fire (>=0.7.0,<0.8.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "omegaconf (>=2.3.0,<3.0.0)",
    "pytorch-lightning (>=2.5.1.post0,<3.0.0)",
    "torch (>=2.7.1,<3.0.0)",
    "torchvision (>=0.22.1,<0.23.0)",
    "mlflow (>=2.22.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"


[tool.isort]
profile = "black"


[tool.flake8]
select = "ANN,C,E,F,W"
ignore = "E203, W503"
max-complexity = 10
max-line-length = 120
